requirements.txt
selenium 
requests
pillow
boto3 